requirements.txt
chardet==3.0.4
click==7.1.2
certifi==2020.6.20
fastjsonschema==2.14.4
idna==2.9
PyYAML==5.3.1
requests==2.24.0
urllib3==1.25.9
